+++ include/ChromaDB/Embeddings/JinaEmbeddingFunction.h
@@ -9,7 +9,7 @@ namespace chromadb {
 	public:
 		JinaEmbeddingFunction(const std::string& apiKey, const std::string& model = "jina-embeddings-v2-base-en", const std::string& baseUrl = "api.jina.ai", const std::string& path = "/v1/embeddings");
 
-		std::vector<std::vector<float>> Generate(const std::vector<std::string>& documents);
+		std::vector<std::vector<double>> Generate(const std::vector<std::string>& documents);
 
 	private:
 		nlohmann::json Request(const nlohmann::json& body) override;
